@@ -1,78 +1,47 @@
-import { useEffect, useState } from "react";
-import { Users } from "./users";
-import "./app.css";
-import Table from "./Table";
-import axios from "axios";
+import { Children } from "react";
+import { createBrowserRouter, RouterProvider, Outlet } from "react-router-dom";
+import Footer from "./components/Footer/Footer";
+import Navbar from "./components/Navbar/Navbar";
+import Home from "./pages/Home/Home";
+import Product from "./pages/Product/Product";
+import Products from "./pages/Products/Products";
+import "./app.scss"
 
-//////////////////////BASIC SEARCH
-
-// function App() {
-//   const [query, setQuery] = useState("");
-//   return (
-//     <div className="app">
-//       <input
-//         className="search"
-//         placeholder="Search..."
-//         onChange={(e) => setQuery(e.target.value.toLowerCase())}
-//       />
-//       <ul className="list">
-//         {Users.filter((asd) =>
-//           asd.first_name.toLowerCase().includes(query)
-//         ).map((user) => (
-//           <li className="listItem" key={user.id}>
-//             {user.first_name}
-//           </li>
-//         ))}
-//       </ul>
-//     </div>
-//   );
-// }
-
-///////////////////////SEARCH ON A DATATABLE
-
-// function App() {
-//   const [query, setQuery] = useState("");
-//   const keys = ["first_name", "last_name", "email"];
-//   const search = (data) => {
-//     return data.filter((item) =>
-//       keys.some((key) => item[key].toLowerCase().includes(query))
-//     );
-//   };
-// return (
-//   <div className="app">
-//       <input
-//         className="search"
-//         placeholder="Search..."
-//         onChange={(e) => setQuery(e.target.value.toLowerCase())}
-//       />
-//     {<Table data={Search(Users)} />}
-//   </div>
-// );
-// }
-
-
-////////////////////// API SEARCH
-
-function App() {
-  const [query, setQuery] = useState("");
-  const [data, setData] = useState([]);
-
-  useEffect(() => {
-    const fetchData = async () => {
-      const res = await axios.get(`http://localhost:5000?q=${query}`);
-      setData(res.data);
-    };
-    if (query.length === 0 || query.length > 2) fetchData();
-  }, [query]);
-
+const Layout = () => {
   return (
     <div className="app">
-        <input
-          className="search"
-          placeholder="Search..."
-          onChange={(e) => setQuery(e.target.value.toLowerCase())}
-        />
-      {<Table data={data} />}
+      <Navbar />
+      <Outlet />
+      <Footer />
+    </div>
+  );
+};
+
+const router = createBrowserRouter([
+  {
+    path: "/",
+    element: <Layout />,
+    children: [
+      {
+        path: "/",
+        element: <Home />,
+      },
+      {
+        path: "/products/:id",
+        element: <Products />,
+      },
+      {
+        path: "/product/:id",
+        element: <Product />,
+      },
+    ],
+  },
+]);
+
+function App() {
+  return (
+    <div>
+      <RouterProvider router={router} />
     </div>
   );
 }
